web/requirements/base.txt
Django
psycopg2
gunicorn
unipath
django-allauth
django-phonenumber-field
wagtail
